// src/app/components/models/profile.ts
export class Profile {
    Id: number;
  	FirstName: string;
    Contact: string;
    Email: string;
    Password: string;
    Company: string;
    LandlordType: string;
    Picture: string;
    Salt: string;
    EmailConfirmed: boolean;
    SecurityStamp: string;
    IsActive: boolean;
    DateCreated: string;
    CreatedBy: string;
    DateModified: string;
    ModifiedBy: string;
    IsDeleted: boolean
}